+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ENHANCEMENTS:
 
 * Added `refresh` argument to `rancher2_catalog` resource
 * Added `name` and `is_external` arguments to `rancher2_user` datasource
+* Added `delete_not_ready_after_secs` and `node_taints` arguments to `node_pool` resource
 
 BUG FIXES: